requirements.txt
# Read me first
#----------------------------------------------------------------
# Please make sure you had already done for these things:
    # Install CUDA and CUDNN
    # Run all of these in Anaconda enviroment
    # install pytorch -version==1.10.0

# Choose the Commend by CUDA version
# CUDA 10.2
    # conda install pytorch==1.10.0 torchvision==0.11.0 torchaudio==0.10.0 cudatoolkit=10.2 -c pytorch
# CUDA 11.3
    # conda install pytorch==1.10.0 torchvision==0.11.0 torchaudio==0.10.0 cudatoolkit=11.3 -c pytorch -c conda-forge
# CPU Only
    # conda install pytorch==1.10.0 torchvision==0.11.0 torchaudio==0.10.0 cpuonly -c pytorch
#----------------------------------------------------------------

# After those things, run this code:
# pip install -r requirements.txt

numpy==1.20.0
opencv-python
Pillow
PyYAML
scipy
tensorboard

beautifulsoup4
requests
html5lib
flask

imutils
pandas
yaml
tqdm
matplotlib
seaborn
psutil
